backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
httpx
gtts
openai
scikit-learn
pandas
numpy
python-dotenv
aiofiles
google-generativeai
Pillow
